+++ src/components/layout/sidebar/sidebar.config.ts
@@ -71,6 +71,8 @@ export const sidebarConfig: SidebarConfig = [
           { type: 'link', label: 'sidebar.items.uiForms', href: routes.ui.forms },
           { type: 'link', label: 'sidebar.items.uiTable', href: routes.ui.table },
           { type: 'link', label: 'sidebar.items.uiDataTable', href: routes.ui.dataTable },
+          { type: 'link', label: 'sidebar.items.uiAreaChart', href: routes.ui.areaChart },
+          { type: 'link', label: 'sidebar.items.uiBarChart', href: routes.ui.barChart },
         ],
       },
       {